Small changes to DB, deleting a user should delete everything related

+++ TravelPalette.DB/Tables/tblLocation.sql
@@ -1,7 +1,7 @@
 CREATE TABLE [dbo].[tblLocation]
 (
 	[Id] INT NOT NULL PRIMARY KEY,
-    [AddressId] INT NOT NULL,
+    [AddressId] BIGINT NOT NULL,
     [Description] VARCHAR(255) NOT NULL,
     [BusinessName] VARCHAR(255) NOT NULL,
     [Coordinates] VARCHAR(50) NOT NULL,
