+++ TravelPalette.UI/Controllers/UserListController.cs
@@ -116,21 +116,20 @@ namespace TravelPalette.UI.Controllers
             }
         }
 
-        public IActionResult AddToList(int id, string businessName, string amenity, double latitude, double longitude)
+        public IActionResult AddToList(double id, string businessName, string amenity, double latitude, double longitude)
         {
             int userId;
             userId = int.Parse(HttpContext.Session.GetString("UserId"));
 
             try
             {
-
-                ListItem listItem = new ListItem() { Id = userId, LocationId = id };
+                ListItem listItem = new ListItem() { Id = userId, LocationId = (int)id };
                 var listItemResult = ListItemManager.Insert(listItem);
                 
                 Location location = new Location()
                 { 
                     Id = -1,
-                    AddressId = id,
+                    AddressId = (int)id,
                     Description = amenity,
                     BusinessName = businessName,
                     Coordinates = latitude + ", " + longitude,
